:art: Rename multistep and progress bar components
