+++ src/components/shared/ProgressBar/index.js
