+++ src/components/shared/ProgressBar/styles.js
