+++ src/components/shared/Stepper/StepProgress.js
