+++ src/components/shared/Stepper/styles.js
