+++ src/modules/Auth/RemoveAccount/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useNavigation } from '@react-navigation/native';
-import MultiStep from 'components/shared/__MultiStep';
+import Stepper from 'components/shared/Stepper';
 import { useAccounts } from 'modules/Accounts/hooks/useAccounts';
 import DeleteAccount from '../DeleteAccount';
 import RemoveAccountConfirmation from '../RemoveAccountConfirmation';
@@ -15,10 +15,10 @@ const RemoveAccount = () => {
   };
 
   return (
-    <MultiStep currentIndex={0}>
+    <Stepper currentIndex={0}>
       <DeleteAccount />
       <RemoveAccountConfirmation onContinue={deleteAccount} />
-    </MultiStep>
+    </Stepper>
   );
 };
