+++ src/modules/Transactions/index.js
@@ -5,7 +5,7 @@ import {
 import connect from 'redux-connect-decorator';
 import { translate } from 'react-i18next';
 
-import MultiStep from 'components/shared/__MultiStep';
+import Stepper from 'components/shared/Stepper';
 import withTheme from 'components/shared/withTheme';
 import { tokenMap } from 'constants/tokens';
 import { B, P } from 'components/shared/toolBox/typography';
@@ -170,7 +170,7 @@ class Send extends React.Component {
     }
 
     return (
-      <MultiStep
+      <Stepper
         ref={(el) => {
           this.nav = el;
         }}
@@ -187,7 +187,7 @@ class Send extends React.Component {
             settings={settings}
           />
         ))}
-      </MultiStep>
+      </Stepper>
     );
   }
 }
